+++ app/models/internal/application/ip.go
@@ -6,7 +6,7 @@ import (
 )
 
 type WhiteIP struct {
-	ID          uint   `gorm:"column:id;primaryKey;unique;not null;comment:id"`
+	ID          int64  `gorm:"column:id;primaryKey;unique;not null;comment:id"`
 	Title       string `gorm:"column:title;type:varchar(50);not null;comment:标题"`
 	IP          string `gorm:"column:ip;type:varchar(50);uniqueIndex;not null;comment:ip"`
 	common.Base `gorm:"embedded"`
