+++ app/models/internal/permission/role_permission.go
@@ -6,8 +6,8 @@ import (
 )
 
 type RolePermissionRelation struct {
-	PermissionID uint `gorm:"column:permission_id;primaryKey;type:bigint;not null;comment:权限id"`
-	RoleID       uint `gorm:"column:role_id;primaryKey;type:bigint;not null;comment:角色id"`
+	PermissionID int64 `gorm:"column:permission_id;primaryKey;type:bigint;not null;comment:权限id"`
+	RoleID       int64 `gorm:"column:role_id;primaryKey;type:bigint;not null;comment:角色id"`
 	common.Base  `gorm:"embedded"`
 }
