+++ app/repositories/permisson_repo.go
@@ -25,18 +25,18 @@ func GetPermissionRepo() PermissionRepo {
 
 type PermissionRepo interface {
 	Create(db *gorm.DB, p *models.Permission) exception.Exception
-	Get(db *gorm.DB, id uint) (*models.Permission, exception.Exception)
+	Get(db *gorm.DB, id int64) (*models.Permission, exception.Exception)
 	GetAll(db *gorm.DB) ([]models.Permission, exception.Exception)
 	GetTop(db *gorm.DB) (*models.Permission, exception.Exception)
-	Update(db *gorm.DB, id uint, param map[string]interface{}) exception.Exception
-	Delete(db *gorm.DB, ids []uint) exception.Exception
+	Update(db *gorm.DB, id int64, param map[string]interface{}) exception.Exception
+	Delete(db *gorm.DB, ids []int64) exception.Exception
 }
 
 func (u *PermissionRepoImpl) Create(db *gorm.DB, p *models.Permission) exception.Exception {
 	return exception.Wrap(response.ExceptionDatabase, db.Create(p).Error)
 }
 
-func (u *PermissionRepoImpl) Get(db *gorm.DB, id uint) (*models.Permission, exception.Exception) {
+func (u *PermissionRepoImpl) Get(db *gorm.DB, id int64) (*models.Permission, exception.Exception) {
 	p := models.Permission{}
 	res := db.Where(&models.Permission{ID: id}).Find(&p)
 	if res.RowsAffected == 0 {
@@ -69,11 +69,11 @@ func (u *PermissionRepoImpl) GetTop(db *gorm.DB) (*models.Permission, exception.
 	return &tp, nil
 }
 
-func (u *PermissionRepoImpl) Update(db *gorm.DB, id uint, param map[string]interface{}) exception.Exception {
+func (u *PermissionRepoImpl) Update(db *gorm.DB, id int64, param map[string]interface{}) exception.Exception {
 	return exception.Wrap(response.ExceptionDatabase,
 		db.Model(&models.Permission{}).Where(&models.Permission{ID: id}).Updates(param).Error)
 }
 
-func (u *PermissionRepoImpl) Delete(db *gorm.DB, ids []uint) exception.Exception {
+func (u *PermissionRepoImpl) Delete(db *gorm.DB, ids []int64) exception.Exception {
 	return exception.Wrap(response.ExceptionDatabase, db.Delete(&models.Permission{}, ids).Error)
 }
